# android/KiraZenbo/src/main/assets/openapi.yaml
openapi: 3.0.0
info:
  title: Zenbo Robot API
  description: A RESTful API to control the Zenbo robot.
  version: 1.0.0
servers:
  - url: /
paths:
  /health:
    get:
      summary: Check the health of the server
      responses:
        '200':
          description: Server is running.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/dialog/speak:
    post:
      summary: Make the robot speak
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: The text for the robot to speak.
              required:
                - text
      responses:
        '200':
          description: Speech command queued successfully.
  /api/dialog/stop_speak:
    post:
      summary: Stop the robot from speaking
      responses:
        '200':
          description: Stop speech command queued successfully.
  /api/dialog/voice_trigger:
    post:
      summary: Enable or disable the voice trigger
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                enable:
                  type: boolean
                  description: true to enable, false to disable.
              required:
                - enable
      responses:
        '200':
          description: Command queued successfully.
  /api/dialog/head_action:
    post:
      summary: Enable or disable the press-on-head action
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                enable:
                  type: boolean
                  description: true to enable, false to disable.
              required:
                - enable
      responses:
        '200':
          description: Command queued successfully.
  /api/face/expression:
    post:
      summary: Set the robot's facial expression
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                expression:
                  type: string
                  description: |
                    The expression to set. Valid options are:
                    ACTIVE, AWARE_LEFT, AWARE_RIGHT, CONFIDENT, DEFAULT, DEFAULT_STILL, 
                    DOUBTING, EXPECTING, HAPPY, HELPLESS, HIDEFACE, IMPATIENT, INNOCENT, 
                    INTERESTED, LAZY, PLEASED, PRETENDING, PROUD, QUESTIONING, SERIOUS, 
                    SHOCKED, SHY, SINGING, TIRED, WORRIED
              required:
                - expression
      responses:
        '200':
          description: Expression command queued successfully.
  /api/face/expression_and_speak:
    post:
      summary: Set the robot's facial expression and speak
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                expression:
                  type: string
                  description: |
                    The expression to set. See /face/expression for valid options.
                text:
                  type: string
                  description: The text for the robot to speak.
              required:
                - expression
                - text
      responses:
        '200':
          description: Command queued successfully.
  /api/utility/follow_face:
    post:
      summary: Ask the robot to follow a detected face
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                enablePreview:
                  type: boolean
                  description: Enable/disable camera preview window for debug.
                  default: false
                largePreview:
                  type: boolean
                  description: Enable/disable full screen preview (only works when enablePreview is true).
                  default: false
      responses:
        '200':
          description: Follow face command queued successfully.
  /api/utility/follow_object:
    post:
      summary: Ask the robot to follow an object
      responses:
        '200':
          description: Follow object command queued successfully.
  /api/utility/track_face:
    post:
      summary: Ask the robot to track a detected face
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                enablePreview:
                  type: boolean
                  description: Enable/disable camera preview window for debug.
                  default: false
                largePreview:
                  type: boolean
                  description: Enable/disable full screen preview (only works when enablePreview is true).
                  default: false
      responses:
        '200':
          description: Track face command queued successfully.
  /api/utility/look_at_user:
    post:
      summary: Ask the robot to look at the user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                doa:
                  type: number
                  format: float
                  description: Direction of arrival (angle) of the user in front of the robot.
              required:
                - doa
      responses:
        '200':
          description: Look-at-user command queued successfully.
  /api/utility/play_action:
    post:
      summary: Perform a predefined motion action
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                number:
                  type: integer
                  description: Action ID to play.
              required:
                - number
      responses:
        '200':
          description: Play action command queued successfully.
  /api/utility/play_emotional_action:
    post:
      summary: Perform a composite emotional action
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                face:
                  type: string
                  description: Robot face expression ID (RobotFace enum name).
                action:
                  type: integer
                  description: Action ID to play.
              required:
                - face
                - action
      responses:
        '200':
          description: Play emotional action command queued successfully.
  /api/utility/get_blue_light_filter_enable:
    get:
      summary: Get whether the screen blue light filter is enabled
      responses:
        '200':
          description: Returns the current blue light filter enable setting.
  /api/utility/get_blue_light_filter_mode:
    get:
      summary: Get the current screen blue light filter mode
      responses:
        '200':
          description: Returns the current blue light filter mode.
  /api/utility/set_blue_light_filter_mode:
    post:
      summary: Set the screen blue light filter mode
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  description: >
                    Blue light filter mode. One of "RdStrong", "Rd03", "Rd02",
                    "Rd01", "RdWeak", "DEF" (DEF disables the filter).
              required:
                - mode
      responses:
        '200':
          description: Set blue light filter mode command queued successfully.
  /api/utility/walk_around:
    post:
      summary: Ask the robot to walk around (deprecated in SDK)
      responses:
        '200':
          description: Walk-around command queued successfully.